Edited function components to only return instead of render and return. Folders added to client folder for actions, components, constat, reducers, and style (including stylesheets).
