+++ client/components/days.jsx
@@ -1,0 +1,15 @@
+import React, { Component } from 'react';
+import { render } from 'react-dom';
+
+
+const Days = props => {
+
+    
+    
+  return (
+    <div>
+      <ul>{Events}</ul>
+    </div>
+  )
+
+}
