+++ client/components/event.jsx
@@ -1,0 +1,13 @@
+import React, { Component } from 'react';
+import { render } from 'react-dom';
+
+const Event = prop => {
+
+
+
+  return (
+    <div>
+      
+    </div>
+  )
+}
