+++ client/components/info.jsx
@@ -1,14 +1,13 @@
 import React, { Component } from 'react';
-import { render } from 'react-dom';
 
 const Info = () => {
-  render () {
+  return (
     <div>
       <p>Year Month</p>
       <button>Algo of the Day!</button>
       <button>Add New Event</button>
     </div>
-  }
+  )
 }
 
 export default Info;
