+++ client/components/main.jsx
@@ -1,0 +1,12 @@
+import React, { Component } from 'react';
+
+const Main = () => {
+  return (
+    <div>
+      <Info />
+      <Month />
+    </div>
+  )
+};
+
+export default Main;
