+++ client/components/planetary.jsx
@@ -1,0 +1,16 @@
+import React, { Component } from 'react';
+import { render } from 'react-dom';
+
+const Planetary = () => {
+  return (
+    <div>
+      <label>SUN</label>
+      <label>MON</label>
+      <label>TUE</label>
+      <label>WED</label>
+      <label>THU</label>
+      <label>FRI</label>
+      <label>SAT</label>
+    </div>
+  )
+}
