--- client/days.jsx
@@ -1,14 +1,0 @@
-const Days = props => {
-
-    
-    
-
-  render() {
-    return (
-      <div>
-        <ul>Events</ul>
-      </div>
-    )
-  }
-
-}
