--- client/main.jsx
@@ -1,15 +1,0 @@
-import React, { Component } from 'react';
-import { render } from 'react-dom';
-
-const Main = () => {
-  render() {
-    return (
-      <div>
-        <Info />
-        <Month />
-      </div>
-    )
-  }
-};
-
-export default Main;
