--- client/planetary.jsx
@@ -1,18 +1,0 @@
-import React, { Component } from 'react';
-import { render } from 'react-dom';
-
-const Planetary = () => {
-  render() {
-    return (
-      <div>
-        <label>SUN</label>
-        <label>MON</label>
-        <label>TUE</label>
-        <label>WED</label>
-        <label>THU</label>
-        <label>FRI</label>
-        <label>SAT</label>
-      </div>
-    )
-  }
-}
