+++ client/style/_variable.scss
@@ -1,0 +1,1 @@
+$lightOrange: #FFCA08;
