+++ client/style/application.scss
@@ -1,0 +1,2 @@
+@import '_variables';
+@import '_style';
