+++ src/store.js
@@ -1,3 +1,6 @@
 import { createStore } from 'redux';
+import { composeWithDevTools } from 'redux-devtools-extention'
+import reducers from '/.reducers/index';
+
